# Crop before resize in Fill