@disintegration 2018-12-15
> This breaks one test, TestFill/Fill_4x4_2x8_Top_Nearest and I'm afraid I'm not sure why

It's because the new algorithm crops the source image to 1x4px first. So it takes a second pixel column and resizes it to 2x8px. The old algorithm provides a more accurate result for such a small image. 
Here's a visual representation using letters in place of pixels.

```
Source image      New result      Old result
                                              
ABCD              BB               BC         
EFGH              BB               BC         
IKLM              FF               FG         
NPQR              FF               FG         
                  KK               KL         
                  KK               KL         
                  PP               PQ         
                  PP               PQ         
```