@disintegration 2018-12-22
Sorry for the delayed response.

Here's another example where the old algorithm does better job. The 3x3 checkers image is scaled to 100x125px. Ideally the individual pixels should be transformed to squares.

```go
func main() {
	src := &image.Gray{
		Rect:   image.Rect(0, 0, 3, 3),
		Stride: 3,
		Pix: []uint8{
			0, 255, 0,
			255, 0, 255,
			0, 255, 0,
		},
	}
	dst := imaging.Fill(src, 100, 125, imaging.TopLeft, imaging.NearestNeighbor)
	imaging.Save(dst, "dst.jpg")
}
```

Old result | New result
---|---
![old](https://user-images.githubusercontent.com/2706839/50374965-14804880-0607-11e9-95ef-041fb6aabdf8.jpg) | ![new](https://user-images.githubusercontent.com/2706839/50374968-1ba75680-0607-11e9-80ed-9af1cfa45b62.jpg)

> Ah I understand, it's because the new algorithm crops so that the pre-resize aspect ratio matches the post-resize aspect ratio. This means that it selects the 1x4 BFKP column because it matches the 0.25 aspect ratio of the destination image.
>
> In order to get the more accurate result, I think we'd have to crop to subpixels before we did the resize. The 1x4 selection would be half column BFKP and half column CGLQ. Upscale these subpixels and, in theory, out pops the correct 2x8 image. In reality, I think we'd have to crop to include any pixel which is fractionally in the selection, resize to the desired height, then crop again to the desired width.
>
> This is getting a bit convoluted for a fairly minor performance gain. Should I try to implement this? I'd be happy to give it a go, but if you think the final code would be too messy for the library then feel free to just close the pull request :)

I think this would work. On the other hand the additional cropping step is the downside of this approach.

After doing some tests, it looks like the new algorithm works fine most of the times. If the size of the original image is not too small, we have enough pixels in both directions to make the initial crop with the desired aspect ratio. We could fallback to the old algorithm for images smaller than, say, 100x100px.

```go
if srcW >= 100 && srcH >= 100 {
    cropAndResize(...)
} else {
    resizeAndCrop(...)
}
```