@bspammer 2018-12-28
I've added the fallback, and updated the documentation. I did try the subpixel method but was still getting inconsistent results, so I think this is a good solution.

I'm guessing the two new functions need tests, it seems to me that the current tests for Fill could also be used for resizeAndCrop as it has identical functionality to the old Fill.

As for cropAndResize, I'm unsure. If it's meant to be used solely for large images then maybe a few golden examples will do?