@disintegration 2018-12-28
Sure, for resizeAndCrop we can use the existing test cases. As for cropAndResize, despite the fact that it's meant to be used with larger images, in tests we just want to ensure that the algorithm works as intended, so I think we still can use a few small-sized test cases. A golden test would be nice to have too.